Round 286
Mr. Kitayuta has kindly given you a string s consisting of lowercase English
letters. You are asked to insert exactly one lowercase English letter into
s to make it a palindrome. A palindrome is a string that reads the same
forward and backward. For example, "noon", "testset" and "a" are all
palindromes, while "test" and "kitayuta" are not.

http://codeforces.com/problemset/problem/505/A

Signed-off-by: Xerof <fercho27@gmail.com>